+++ src/main/java/com/railway/reservation_system/payment/PaymentStrategy.java
@@ -4,6 +4,6 @@ import com.railway.reservation_system.exception.PaymentException;
 
 public interface PaymentStrategy {
     
-    public boolean pay(int requestedAmount) throws PaymentException;
+    public boolean pay(float requestedAmount) throws PaymentException;
 
 }
